backend/requirements.txt
# =====================================================
# Beyond Words — Backend Requirements (Enhanced)
# =====================================================

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Processing & ML
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
pandas>=2.0.0
joblib>=1.3.0

# Deep Learning
torch>=2.0.0
tensorflow>=2.15.0  # For CNN/CRNN models
keras>=2.15.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0
audioread>=3.0.0
resampy>=0.4.0

# NLP & Speech Recognition
transformers>=4.35.0
openai-whisper>=20231117
sentencepiece>=0.1.99
protobuf>=4.25.0
torchaudio>=2.0.0  # For Wav2Vec2/HuBERT

# Database Support
# PostgreSQL
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0  # Database migrations

# MongoDB
pymongo==4.6.0
motor==3.3.2  # Async MongoDB driver

# Model Evaluation & Visualization
matplotlib==3.8.2
seaborn==0.13.0
scikit-plot==0.3.7

# Utilities
python-dotenv==1.0.0  # Environment variables
pydantic==2.5.0
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4  # Password hashing

# Logging & Monitoring
loguru==0.7.2

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # For testing async endpoints
